+++ student-store-ui/src/components/About/About.css
@@ -1,32 +1,32 @@
 .about {
-    margin-top: 70px;
+  margin-top: 70px;
 }
 
 .about-box {
-    display: flex;
-    justify-content: left;
-    flex: 2;
-    background: #2c4047;
-    border: solid 1px #e0e0e0
-  }
+  display: flex;
+  justify-content: left;
+  flex: 2;
+  background: #2c4047;
+  border: solid 1px #e0e0e0
+}
   
 .about-text {
-    display: flex;
-    flex-direction: column;
-    justify-content: left;
-    flex: 1;
-    background-color: #fff;
-    padding: 28px 36px;
-  }
-  
-  .about-logo {
-    display: flex;
-    flex-direction: row;
-    height: 90%;
-    padding: 5% 0% 5% 0%;
-  }
+  display: flex;
+  flex-direction: column;
+  justify-content: left;
+  background-color: #fff;
+  flex: 1;
+  padding: 28px 36px;
+}
 
-  .about-logo img{
-    height: 200px;
-    width: 200px;
-  }
+.about-logo {
+  display: flex;
+  flex-direction: row;
+  height: 90%;
+  padding: 5% 0% 5% 0%;
+}
+
+.about-logo img{
+  height: 200px;
+  width: 200px;
+}
