+++ student-store-ui/src/components/CategoryBar/CategoryBar.jsx
@@ -1,19 +1,22 @@
-import * as React from "react"
 import "./CategoryBar.css"
+
 import Chip from "../Chip/Chip"
 
-export default function CategoryBar({ categories, category, setCategory }) {
+export default function CategoryBar({ category, setCategory }) {
+  const categories = ["All Categories", "Food", "Clothing", "Tech", "Accessories"]
+
   return (
     <nav className="category-bar">
-        {categories?.map((label, index) => {
-            return (
-            <Chip
-                category={label}
-                isActive={category === label}
-                handleClick={() => setCategory(label)}
-                key={String(index).concat("category")}
-            />)
-        })}
+      {categories?.map((label, index) => {
+        return (
+          <Chip
+            category={label}
+            isActive={category === label}
+            handleClick={() => setCategory(label)}
+            key={String(index).concat("category")}
+          />
+        )
+      })}
     </nav>
   )
 }
