+++ student-store-ui/src/components/Contact/Contact.css
@@ -1,7 +1,6 @@
-
 .contact {
     margin-top: 70px;
-  }
+}
 
 .contact-text-box {
     display: flex;
@@ -18,6 +17,7 @@
     margin-bottom: 20px;
 
 }
+
 .contact-text {
     width: 100px;
 }
