+++ student-store-ui/src/components/Footer/Footer.jsx
@@ -1,52 +1,52 @@
 import "./Footer.css"
 
 export default function Footer() {
-    return (
-        <section className="footer">
-              <div className="footer-items-box">
-                <section className="footer-item">
-                  <h4>Categories</h4>
-                  <div>All Categories</div>
-                  <div>Clothing</div>
-                  <div>Food</div>
-                  <div>Accessories</div>
-                  <div>Tech</div>
-                </section>
-                <section className="footer-item">
-                  <h4>Company</h4>
-                  <div>About Us</div>
-                  <div>Find a Store</div>
-                  <div>Terms</div>
-                  <div>Sitemap</div>
-                  <div>Careers</div>
-                </section>
-                <section className="footer-item">
-                  <h4>Support</h4>
-                  <div>Contact Us</div>
-                  <div>Money Refund</div>
-                  <div>Order Status</div>
-                  <div>Shipping Info</div>
-                  <div>Open Dispute</div>
-                </section>
-                <section className="footer-item">
-                  <h4>Account</h4>
-                  <div>Login</div>
-                  <div>Register</div>
-                  <div>Account Setting</div>
-                  <div>My Orders</div>
-                </section>
-                <section className="footer-item">
-                  <h4>Socials</h4>
-                  <div>Facebook</div>
-                  <div>Twitter</div>
-                  <div>LinkedIn</div>
-                  <div>Instagram</div>
-                  <div>Youtube</div>
-                </section>
-                <section className="footer-item">
-                  <h4>Our App</h4>
-                </section>
-              </div>
-            </section>
-    )
+  return (
+      <section className="footer">
+        <div className="footer-items-box">
+          <section className="footer-item">
+            <h4>Categories</h4>
+            <div>All Categories</div>
+            <div>Clothing</div>
+            <div>Food</div>
+            <div>Accessories</div>
+            <div>Tech</div>
+          </section>
+          <section className="footer-item">
+            <h4>Company</h4>
+            <div>About Us</div>
+            <div>Find a Store</div>
+            <div>Terms</div>
+            <div>Sitemap</div>
+            <div>Careers</div>
+          </section>
+          <section className="footer-item">
+            <h4>Support</h4>
+            <div>Contact Us</div>
+            <div>Money Refund</div>
+            <div>Order Status</div>
+            <div>Shipping Info</div>
+            <div>Open Dispute</div>
+          </section>
+          <section className="footer-item">
+            <h4>Account</h4>
+            <div>Login</div>
+            <div>Register</div>
+            <div>Account Setting</div>
+            <div>My Orders</div>
+          </section>
+          <section className="footer-item">
+            <h4>Socials</h4>
+            <div>Facebook</div>
+            <div>Twitter</div>
+            <div>LinkedIn</div>
+            <div>Instagram</div>
+            <div>Youtube</div>
+          </section>
+          <section className="footer-item">
+            <h4>Our App</h4>
+          </section>
+        </div>
+    </section>
+  )
 }
