--- student-store-ui/src/components/Logo/Logo.css
@@ -1,3 +1,0 @@
-.logo {
-    color: #fff;
-}
