+++ student-store-ui/src/components/Logo/Logo.jsx
@@ -1,8 +1,5 @@
-import * as React from "react"
-import "./Logo.css"
-
 export default function Logo() {
   return (
-      <img src="https://codepath-student-store-demo.surge.sh/assets/codepath.f1b3e41a.svg"></img>
+    <img src="https://codepath-student-store-demo.surge.sh/assets/codepath.f1b3e41a.svg" />
   )
 }
