+++ student-store-ui/src/components/ProductCard/ProductCard.css
@@ -6,15 +6,15 @@
 }
 
 .product-card img {
+    object-fit: cover;
     width: 100%;
     height: 100%;
     max-height: 400px;
     min-height: 100px;
-    object-fit: cover;
 }
 
 .info-box {
-    padding: 5px 5px;
     background-color: #fff;
-    border: solid 1px #e0e0e0
+    border: solid 1px #e0e0e0;
+    padding: 5px 5px;
 }
