+++ student-store-ui/src/components/ProductDetails/ProductDetails.css
@@ -1,22 +1,22 @@
 .product-card-description {
-    padding-top: 5%;
     display: flex;
     flex-direction: column;
     justify-content: center;
     align-items: center;
     width: 100%;
     height: 100%;
+    padding-top: 5%;
 }
 
 .product-card-description img {
-    max-width: 300px;
     object-fit: cover;
     justify-content: center;
     padding-bottom: 5%;
+    max-width: 300px;
 }
 
 .info-box {
+    border: solid 1px #e0e0e0;
     padding: 5px 5px;
     background-color: #fff;
-    border: solid 1px #e0e0e0
 }
