+++ student-store-ui/src/components/Sidebar/Sidebar.jsx
@@ -1,4 +1,3 @@
-import * as React from "react"
 import "./Sidebar.css"
 
 export default function Sidebar() {
