--- student-store-ui/src/components/Subnavbar/Subnavbar.css
@@ -1,10 +1,0 @@
-.subnav-wrap {
-    display: flex;
-    justify-content: center;
-}
-
-.subnavbar {
-    display: flex;
-    flex-direction: column;
-    color: #00c385;
-}
